+++ app/helpers/reviews_helper.rb
@@ -1,0 +1,2 @@
+module ReviewsHelper
+end
