+++ app/models/developer.rb
