+++ app/models/publisher.rb
