+++ db/migrate/20230301130932_add_user_to_reviews.rb
@@ -1,0 +1,5 @@
+class AddUserToReviews < ActiveRecord::Migration[7.0]
+  def change
+    add_reference :reviews, :user, null: false, foreign_key: true
+  end
+end
