+++ db/migrate/20230301141016_add_nickname_to_users.rb
@@ -1,0 +1,5 @@
+class AddNicknameToUsers < ActiveRecord::Migration[7.0]
+  def change
+    add_column :users, :nickname, :string
+  end
+end
