+++ test/models/genre_test.rb
